dsp: Split DSPState structure off into new file
